move old script
