+++ examples/speaker_tasks/recognition/speaker_reco_finetune.py
@@ -12,83 +12,38 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""
-This script demonstrates how to use pretrained .nemo speaker recognition model to finetune on your domain set
-Usage:
-python speaker_reco_finetune.py --pretrained_model='/path/to/.nemo_or_.ckpt/file' --finetune_config_file=/path/to/finetune/config/yaml/file'
-
-in finetune config file make sure to change
-- train manifest
-- validation manifest
-- decoder num classes
-- learning rate
-- num of epochs
-
-for finetuning tips see: https://github.com/NVIDIA/NeMo/blob/main/tutorials/speaker_tasks/Speaker_Recognition_Verification.ipynb
-"""
-
-
-from argparse import ArgumentParser
+import os
+import torch.nn as nn
 
 import pytorch_lightning as pl
 from omegaconf import OmegaConf
+from pytorch_lightning import seed_everything
 
 from nemo.collections.asr.models import EncDecSpeakerLabelModel
+from nemo.core.config import hydra_runner
 from nemo.utils import logging
 from nemo.utils.exp_manager import exp_manager
 
+seed_everything(42)
 
-def main():
-    parser = ArgumentParser()
-    parser.add_argument(
-        "--pretrained_model", type=str, default="titanet_large", required=False, help="Pass your trained .nemo model",
-    )
-    parser.add_argument(
-        "--finetune_config_file",
-        type=str,
-        required=True,
-        help="path to speakernet config yaml file to load train, validation dataset and also for trainer parameters",
-    )
 
-    parser.add_argument(
-        "--freeze_encoder",
-        type=bool,
-        required=False,
-        default=True,
-        help="True if speakernet encoder paramteres needs to be frozen while finetuning",
-    )
+@hydra_runner(config_path="conf", config_name="SpeakerNet_verification_3x2x256.yaml")
+def main(cfg):
 
-    args = parser.parse_args()
+    logging.info(f'Hydra config: {OmegaConf.to_yaml(cfg)}')
+    trainer = pl.Trainer(**cfg.trainer)
+    log_dir = exp_manager(trainer, cfg.get("exp_manager", None))
+    speaker_model = EncDecSpeakerLabelModel(cfg=cfg.model, trainer=trainer)
+    pretrained = EncDecSpeakerLabelModel.from_pretrained("titanet_large")
+    pretrained.decoder.final = speaker_model.decoder.final
+    speaker_model.load_state_dict(pretrained.state_dict())
+    # speaker_model.maybe_init_from_pretrained_checkpoint(cfg)
+    trainer.fit(speaker_model)
 
-    if args.pretrained_model.endswith('.nemo'):
-        logging.info(f"Using local speaker model from {args.pretrained_model}")
-        speaker_model = EncDecSpeakerLabelModel.restore_from(restore_path=args.pretrained_model)
-    elif args.pretrained_model.endswith('.ckpt'):
-        logging.info(f"Using local speaker model from checkpoint {args.pretrained_model}")
-        speaker_model = EncDecSpeakerLabelModel.load_from_checkpoint(checkpoint_path=args.pretrained_model)
-    else:
-        logging.info("Using pretrained speaker recognition model from NGC")
-        speaker_model = EncDecSpeakerLabelModel.from_pretrained(model_name=args.pretrained_model)
-
-    finetune_config = OmegaConf.load(args.finetune_config_file)
-
-    if 'test_ds' in finetune_config.model and finetune_config.model.test_ds is not None:
-        finetune_config.model.test_ds = None
-        logging.warning("Removing test ds")
-
-    speaker_model.setup_finetune_model(finetune_config.model)
-    finetune_trainer = pl.Trainer(**finetune_config.trainer)
-    speaker_model.set_trainer(finetune_trainer)
-
-    _ = exp_manager(finetune_trainer, finetune_config.get('exp_manager', None))
-    speaker_model.setup_optimization(finetune_config.optim)
-
-    if args.freeze_encoder:
-        for param in speaker_model.encoder.parameters():
-            param.requires_grad = False
-
-    finetune_trainer.fit(speaker_model)
-
+    if hasattr(cfg.model, 'test_ds') and cfg.model.test_ds.manifest_filepath is not None:
+        trainer = pl.Trainer(devices=1, accelerator=cfg.trainer.accelerator)
+        if speaker_model.prepare_test(trainer):
+            trainer.test(speaker_model)
 
 if __name__ == '__main__':
     main()
