--- examples/speaker_tasks/recognition/speaker_reco_ft.py
@@ -1,49 +1,0 @@
-# Copyright (c) 2020, NVIDIA CORPORATION.  All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import os
-import torch.nn as nn
-
-import pytorch_lightning as pl
-from omegaconf import OmegaConf
-from pytorch_lightning import seed_everything
-
-from nemo.collections.asr.models import EncDecSpeakerLabelModel
-from nemo.core.config import hydra_runner
-from nemo.utils import logging
-from nemo.utils.exp_manager import exp_manager
-
-seed_everything(42)
-
-
-@hydra_runner(config_path="conf", config_name="SpeakerNet_verification_3x2x256.yaml")
-def main(cfg):
-
-    logging.info(f'Hydra config: {OmegaConf.to_yaml(cfg)}')
-    trainer = pl.Trainer(**cfg.trainer)
-    log_dir = exp_manager(trainer, cfg.get("exp_manager", None))
-    speaker_model = EncDecSpeakerLabelModel(cfg=cfg.model, trainer=trainer)
-    pretrained = EncDecSpeakerLabelModel.from_pretrained("titanet_large")
-    pretrained.decoder.final = speaker_model.decoder.final
-    speaker_model.load_state_dict(pretrained.state_dict())
-    # speaker_model.maybe_init_from_pretrained_checkpoint(cfg)
-    trainer.fit(speaker_model)
-
-    if hasattr(cfg.model, 'test_ds') and cfg.model.test_ds.manifest_filepath is not None:
-        trainer = pl.Trainer(devices=1, accelerator=cfg.trainer.accelerator)
-        if speaker_model.prepare_test(trainer):
-            trainer.test(speaker_model)
-
-if __name__ == '__main__':
-    main()
